Role: component refactor
