+++ lib/frontend/src/actions/roles.js
@@ -345,3 +345,14 @@ export const updateItem = (itemId, data) => async dispatch => {
     )
   }
 }
+
+export const dispatchForm = itemType => dispatch => {
+  switch (itemType) {
+    case 'indicator':
+      dispatch(indicatorForm())
+      break
+    case 'checklist':
+      dispatch(checkForm())
+      break
+  }
+}
